Up to x 1.5 speed up for finding value types

* Improve Dictionary FindEntry CQ https://github.com/dotnet/coreclr/pull/15460
* Improve Dictionary TryInsert CQ https://github.com/dotnet/coreclr/pull/15462
* Use `EqualityComparer<TKey>.Default` (only from #15411)
* 1-bases `_buckets` so the don't be set to `-1` on each resize

Adds 5kB (0.16% of base) to `System.Private.CoreLib.ni.dll` https://github.com/dotnet/coreclr/pull/15419#issuecomment-353247960

Before
```
                                Method |
-------------------------------------- |
    ContainsKeySimplePositive_IntKey   | IEqualityComparer
    ContainsKeySimpleNegative_IntKey   | IEqualityComparer
 ContainsKeySimplePositive_StringKey   | IEqualityComparer
 ContainsKeySimpleNegative_StringKey   | IEqualityComparer
   ContainsKeySimplePositive_ObjectKey | IEqualityComparer
   ContainsKeySimpleNegative_ObjectKey | IEqualityComparer
```
After
```
                                Method |
-------------------------------------- |
    ContainsKeySimplePositive_IntKey   | devirtualized
    ContainsKeySimpleNegative_IntKey   | devirtualized
 ContainsKeySimplePositive_StringKey   | IEqualityComparer
 ContainsKeySimpleNegative_StringKey   | IEqualityComparer
   ContainsKeySimplePositive_ObjectKey | virtual
   ContainsKeySimpleNegative_ObjectKey | virtual
```

```
                          Method |Items |       Current |   Single func | Change |
-------------------------------- |----- |--------------:|--------------:|-------:|
   ContainsKey_Single_Yes_IntKey | 1000 |     11.231 ns |      7.110 ns | x 1.58 |
    ContainsKey_Single_No_IntKey | 1000 |     10.794 ns |      7.179 ns | x 1.50 |
          ContainsKey_All_IntKey | 1000 | 10,815.405 ns |  7,220.128 ns | x 1.50 |
ContainsKey_Single_Yes_StringKey | 1000 |     20.489 ns |     19.925 ns | x 1.03 |
 ContainsKey_Single_No_StringKey | 1000 |     12.747 ns |     13.414 ns | x 0.95 |
ContainsKey_Single_Yes_ObjectKey | 1000 |     11.233 ns |      9.119 ns | x 1.23 |
 ContainsKey_Single_No_ObjectKey | 1000 |     11.192 ns |      8.560 ns | x 1.31 |
                      Add_IntKey | 1000 | 27,573.689 ns | 22,503.652 ns | x 1.23 |
                   Add_StringKey | 1000 | 76,095.748 ns | 77,253.701 ns | x 0.99 |
                   Add_ObjectKey | 1000 | 68,845.402 ns | 67,529.550 ns | x 1.02 |
              PreSizedAdd_IntKey | 1000 | 11,558.450 ns |  9,658.953 ns | x 1.20 |
           PreSizedAdd_StringKey | 1000 | 57,360.936 ns | 57,710.254 ns | x 0.99 |
           PreSizedAdd_ObjectKey | 1000 | 54,780.942 ns | 50,300.463 ns | x 1.09 |
```

Performance-wise, multipath was faster but that increased code-gen by 27kB https://github.com/dotnet/coreclr/pull/15419#issuecomment-352159250

Resolves: https://github.com/dotnet/coreclr/issues/16258