@benaadams 2017-12-07
if `comparer` not default (null) use that
if type is struct `EqualityComparer<TKey>.Default.Equals`
if type object use `key.Equals(otherKey)` (key is never null)

Would box with a nullable key? However, you can't have a null key, so that seems like an antipattern anyway?