@jkotas 2017-12-07
@dotnet-bot test Windows_NT x64 Checked corefx_baseline