@jkotas 2017-12-07
@dotnet-bot test Ubuntu x64 Checked corefx_baseline