@jkotas 2017-12-07
Do you have any numbers for this? Throughput, code size?