@benaadams 2017-12-07
Was looking for this https://github.com/dotnet/corefx/blob/master/Documentation/project-docs/benchmarking.md should have them shortly