@omariom 2017-12-07
Is there a way to run perf tests from CoreFX? There tests for Dictionary and HashSet there.