@benaadams 2017-12-07
I've never particularly worked out how to run the Microsoft.Xunit.Performance tests