@danmosemsft 2017-12-07
I haven't done it in a while, but [this ](https://github.com/dotnet/corefx/blob/master/Documentation/project-docs/performance-tests.md) describes how to run the tests and of course you could build with `/p:CoreCLROverridePath` before doing that.