@benaadams 2017-12-07
Failures are in Regex tests
```
System.OutOfMemoryException : Exception of type 'System.OutOfMemoryException' was thrown.
```