@benaadams 2017-12-08
While its an improvement; its a bit fragile (changing the code a little can loose the improvement) - so I think it needs work
```
             Method |     Mean |    Error |   StdDev |
------------------- |---------:|---------:|---------:|
 Knucleotide (Pre)  | 72.88 ms | 1.493 ms | 1.777 ms |
 Knucleotide (Post) | 67.64 ms | 1.349 ms | 1.847 ms |
```

The corefx tests produce odd results; with variations in the same tests at different sizes; feels like GC - also there are no object key tests (are string keys, but that's a special case)

I think this might be a longer journey...