@benaadams 2017-12-08
> Did you use BDN?

Yeah, `[InProcess]` publish self-contained; copy coreclr release tests bin dlls (corefx+coreclr) on top; duplicate folder for each variant; copy Product coreclr dll over for each variant - bit manual; but means don't have to rebuild corefx each time to produce a NetCoreApp package using the right coreclr; which is a very slow iteration step