@omariom 2017-12-08
@benaadams  
What is `Knucleotide`? 
I  started porting Perf.Dictionary.cs but stuck at set_Item - there was a mysterious degradation, though with manual nano-benchmarking there was improvement. 
I will publish them as gist tomorrow.