@benaadams 2017-12-08
> What is Knucleotide?

Its interesting use of Dictionary from [benchmarks game](http://benchmarksgame.alioth.debian.org/u64q/knucleotide-description.html#knucleotide), rather than just testing raw function time; though not specifically just testing dictionary 

Example implementations in the [JIT Performance/CodeQuality](https://github.com/dotnet/coreclr/tree/master/tests/src/JIT/Performance/CodeQuality/BenchmarksGame/k-nucleotide) directory; though I think I was using an easier iteration of the algorithm.