@benaadams 2017-12-15
@omariom tests (after adding `[Benchmark(OperationsPerInvoke = Iterations)]` to `ContainsKey_Int_True`)

Before
```
               Method | Size_Common |       Mean |     Error |     StdDev |     Median |
--------------------- |------------ |-----------:|----------:|-----------:|-----------:|
                  Add |        1000 | 56.1203 ns | 7.3247 ns | 21.5972 ns | 61.6102 ns |
              GetItem |        1000 | 10.8154 ns | 0.2153 ns |  0.3827 ns | 10.7159 ns |
              SetItem |        1000 | 12.0262 ns | 0.2308 ns |  0.2565 ns | 11.9275 ns |
          TryGetValue |        1000 | 12.4566 ns | 0.2480 ns |  0.5285 ns | 12.3705 ns |
 ContainsKey_Int_True |        1000 |  0.8052 ns | 0.1351 ns |  0.3919 ns |  0.5448 ns |
                  Add |       10000 | 76.9773 ns | 8.5637 ns | 25.2502 ns | 75.0018 ns |
              GetItem |       10000 | 10.9235 ns | 0.2289 ns |  0.4676 ns | 10.7257 ns |
              SetItem |       10000 | 12.0162 ns | 0.2401 ns |  0.4509 ns | 11.8815 ns |
          TryGetValue |       10000 | 12.3240 ns | 0.2453 ns |  0.4955 ns | 12.1962 ns |
 ContainsKey_Int_True |       10000 |  5.4632 ns | 0.1880 ns |  0.4919 ns |  5.1851 ns |
                  Add |      100000 | 90.1326 ns | 9.9792 ns | 29.4238 ns | 82.4575 ns |
              GetItem |      100000 | 10.7685 ns | 0.2125 ns |  0.4528 ns | 10.6014 ns |
              SetItem |      100000 | 11.9630 ns | 0.2374 ns |  0.5780 ns | 11.6291 ns |
          TryGetValue |      100000 | 12.4327 ns | 0.2850 ns |  0.4602 ns | 12.3095 ns |
 ContainsKey_Int_True |      100000 | 64.8813 ns | 6.9787 ns | 20.4674 ns | 52.6575 ns |
```
After
```
               Method | Size_Common |       Mean |     Error |     StdDev |     Median |
--------------------- |------------ |-----------:|----------:|-----------:|-----------:|
                  Add |        1000 | 50.4612 ns | 6.9421 ns | 20.4689 ns | 50.4973 ns |
              GetItem |        1000 |  7.3103 ns | 0.1430 ns |  0.1469 ns |  7.3099 ns |
              SetItem |        1000 |  7.9682 ns | 0.1664 ns |  0.1557 ns |  7.9311 ns |
          TryGetValue |        1000 |  8.1403 ns | 0.1695 ns |  0.3613 ns |  8.0343 ns |
 ContainsKey_Int_True |        1000 |  0.4659 ns | 0.1413 ns |  0.4053 ns |  0.2477 ns |
                  Add |       10000 | 74.7613 ns | 8.3463 ns | 24.6093 ns | 70.2436 ns |
              GetItem |       10000 |  7.3778 ns | 0.1382 ns |  0.1225 ns |  7.3761 ns |
              SetItem |       10000 |  8.3528 ns | 0.2027 ns |  0.5849 ns |  8.0981 ns |
          TryGetValue |       10000 |  7.9902 ns | 0.1567 ns |  0.2145 ns |  7.9654 ns |
 ContainsKey_Int_True |       10000 |  2.7554 ns | 0.2278 ns |  0.5839 ns |  2.5304 ns |
                  Add |      100000 | 85.5660 ns | 9.0746 ns | 26.7566 ns | 74.2813 ns |
              GetItem |      100000 |  7.3365 ns | 0.1419 ns |  0.1894 ns |  7.2229 ns |
              SetItem |      100000 |  7.8615 ns | 0.1997 ns |  0.1770 ns |  7.8193 ns |
          TryGetValue |      100000 |  7.9205 ns | 0.1552 ns |  0.1963 ns |  7.8259 ns |
 ContainsKey_Int_True |      100000 | 42.7261 ns | 8.8222 ns | 25.8741 ns | 25.8229 ns |
```