@benaadams 2017-12-21
1-basing the buckets brings the code size down a little
```
otal bytes of diff: 27150 (0.80% of base)
    diff is a regression.

Total byte diff includes 22018 bytes from reconciling methods
        Base had    0 unique methods,        0 unique bytes
        Diff had   24 unique methods,    22018 unique bytes

Top file regressions by size (bytes):
       27150 : System.Private.CoreLib.dasm (0.80% of base)

1 total files with size differences (0 improved, 1 regressed), 0 unchanged.

Top method regessions by size (bytes):
        3171 : System.Private.CoreLib.dasm - Dictionary`2:.ctor(ref,ref):this (58 methods)
        2619 : System.Private.CoreLib.dasm - Dictionary`2:FindEntryDefaultComparer(ref):int:this (0/11 methods)
        2587 : System.Private.CoreLib.dasm - Dictionary`2:TryInsertDefaultComparer(ref,struct,ubyte):bool:this (0/4 methods)
        1953 : System.Private.CoreLib.dasm - Dictionary`2:TryInsertDefaultComparer(struct,ref,ubyte):bool:this (0/3 methods)
        1872 : System.Private.CoreLib.dasm - Dictionary`2:TryInsertDefaultComparer(ref,int,ubyte):bool:this (0/3 methods)

Top method improvements by size (bytes):
       -3964 : System.Private.CoreLib.dasm - Dictionary`2:Resize(int,bool):this (29 methods)
       -1598 : System.Private.CoreLib.dasm - Dictionary`2:.ctor(int,ref):this (29 methods)
        -783 : System.Private.CoreLib.dasm - Dictionary`2:Initialize(int):this (29 methods)
        -768 : System.Private.CoreLib.dasm - CultureData:get_RegionNames():ref
        -650 : System.Private.CoreLib.dasm - Dictionary`2:FindEntry(ref):int:this (11 methods)

219 total methods with size differences (75 improved, 144 regressed), 16549 unchanged.
```
Performance is still good
```
                           Method |   Items |            Current |                New |  Change |
--------------------------------- |-------- |-------------------:|-------------------:|--------:|
    ContainsKey_Single_Yes_IntKey |    1000 |          11.231 ns |           6.235 ns |  x 1.80 |
     ContainsKey_Single_No_IntKey |    1000 |          10.794 ns |           5.986 ns |  x 1.80 |
           ContainsKey_All_IntKey |    1000 |      10,815.405 ns |       5,968.820 ns |  x 1.81 |
 ContainsKey_Single_Yes_StringKey |    1000 |          20.489 ns |          21.079 ns |  x 0.97 |
  ContainsKey_Single_No_StringKey |    1000 |          12.747 ns |          14.186 ns |  x 0.90 |
 ContainsKey_Single_Yes_ObjectKey |    1000 |          11.233 ns |           9.129 ns |  x 1.23 |
  ContainsKey_Single_No_ObjectKey |    1000 |          11.192 ns |           9.103 ns |  x 1.23 |
                       Add_IntKey |    1000 |      27,573.689 ns |      22,362.911 ns |  x 1.23 |
                    Add_StringKey |    1000 |      76,095.748 ns |      75,990.481 ns |  x 1.00 |
                    Add_ObjectKey |    1000 |      68,845.402 ns |      67,466.579 ns |  x 1.02 |
               PreSizedAdd_IntKey |    1000 |      11,558.450 ns |       9,490.586 ns |  x 1.22 |
            PreSizedAdd_StringKey |    1000 |      57,360.936 ns |      57,180.370 ns |  x 1.00 |
            PreSizedAdd_ObjectKey |    1000 |      54,780.942 ns |      49,087.879 ns |  x 1.12 |
```